==> 01-prepare-dataframe.ipynb <==
# Prepare Anesthetized DataFrame

```
Here, we will prepare a .csv dataframe from the h5files in the original dataset.

Notes
-----
1. Electrode locations are not part of the CRCNS aa-4 dataset and must be acquired separately.
2. We are taking {field L/CML complex} as A1 and {NCM, CMM} as A2 (Robotka et al 2023).
3. For WhiWhi4522M, Site5 uses different stimuli than other sites. Recordings from this site have been dropped.
4. Noise stimuli have been dropped.
```

## Import Modules

import numpy as np
import pandas as pd
import h5py
import os

import warnings
warnings.filterwarnings("ignore")

## Utility Functions

```
In the dataset, there are "raw" electrode files, and spike-sorted single-unit files. We only want to use the spike-sorted files. Electrode location files are not part of the original dataset and must be placed separately in the 'data/anesthetized' directory.
```

def isSpikeSorted(filename):
    if filename.endswith('.h5') and '_ss' in filename:
        return True
    return False

def isLocFile(filename):
    if filename.endswith('.txt') and 'Histo' in filename:
        return True
    return False

## Create DataFrame

### Collect Data

data_root = os.path.join('..', '..', 'data', 'anesthetized')
birds = [d for d in os.scandir(data_root) if os.path.isdir(d)]
birds

dict_list = []    # list of data dicts
# loop through subjects
print("Collecting data...")
for bird in birds:
    print("Currently working on: ", bird.name)
    # loop through data files
    for data_file in os.scandir(bird.path):
        # skip over raw electrode files
        if not isSpikeSorted(data_file.name):
            continue
        
        # read data_file for a given single_unit
        single_unit = h5py.File(data_file, 'r')
        main_key = list(single_unit.keys())[0]
        assert 'Call' in main_key
        
        # collect single_unit data for all stimuli
        stim_numbers = single_unit[main_key].keys()
        for stim_number in stim_numbers:
            attributes = single_unit[main_key][stim_number].attrs.keys()
            # loop through attributes
            data_dict = {}
            data_dict['subject'] = bird.name
            data_dict['site'] = str(single_unit.attrs['site'], 'utf-8')
            data_dict['electrode'] = data_file.name[:-7] + '.h5'
            data_dict['single_unit'] = data_file.name[:-3]
            data_dict['sortType'] = str(single_unit.attrs['sortType'], 'utf-8')
            data_dict['stim_number'] = stim_number
            
            for attr in attributes:
                data_dict[attr] = single_unit[main_key][stim_number].attrs[attr]
                if not isinstance(data_dict[attr], float):
                    try:
                        # string types are stored as byte-strings in the dataset
                        data_dict[attr] = str(single_unit[main_key][stim_number].attrs[attr], 'utf-8')
                    except:
                        pass

            wavefile = str(single_unit[main_key][stim_number].attrs['original_wavfile'], 'utf-8')
            data_dict['original_wavfile'] = wavefile[wavefile.rindex('/')+1:]
            
            dict_list.append(data_dict)
                
print("Finished!")

df = pd.DataFrame(dict_list).drop_duplicates()
df.sample(5)

### Add Hemisphere Data

df_list = []
for file in os.scandir(data_root):
    if not isLocFile(file.name):
        continue
    data_frame = pd.read_csv(file.path, delim_whitespace=True,
                          names=['electrode', 'brain_region', 'x?', 'y?', 'hemisphere', 'unknown?'])
    data_frame = data_frame[['electrode', 'brain_region', 'hemisphere']]
    data_frame['subject'] = file.name[13:24]
    df_list.append(data_frame)

loc_df = pd.concat(df_list)
loc_df.sample(5)

df = df.merge(loc_df, on=['subject', 'electrode'], how='left')
del df['electrode']
df.sample(5)

df['brain_region'].isnull().sum()/len(df['brain_region'])

## Clean DataFrame
### Drop Unnecessary Columns

df.columns

df = df[['subject', 'site', 'single_unit', 'sortType', 'stim_number',
        'original_wavfile', 'callerAge', 'callid', 'stim_duration', 'stim_source', 
        'stim_source_sex', 'stim_type', 'brain_region', 'hemisphere']]
df.sample(5)

## Preprocess DataFrame

### vocid = callid + song

df['vocid'] = np.where(df['stim_type']=='call', df['callid'], df['stim_type'])    # where True yield x, otherwise y
df = df[df.vocid != 'mlnoise']
df.sample(5)

df['vocid'].value_counts()

### Filter Noise and Multi-Units

df['sortType'].value_counts()

df['sortType'].unique()[3]

df['sortType'] = df['sortType'].replace({df['sortType'].unique()[3]: 'single'}, regex=True)

df = df[(df['sortType']=='single')]
df.sample(5)

df['sortType'].value_counts()

### Add Column for A1/A2

df['brain_region'].unique()

'''
Field L regions L1, L2, L3, and lateral caudal mesopallium (CM) form the primary auditory cortex in birds (13, 14). 
Field L2a and L2b are the intermediate (thalamorecipient) regions. 
Field L1 and CML are the superficial regions, and field L3 is the deep region. 
Caudal nidopallium (NC) is a secondary auditory area. 
Area L (13) has cytoarchitecture similar to L2b but is not known to receive thalamic input. 

Reference:
Coding principles of the canonical cortical microcircuitin the avian brain
Ana Calabreseaand Sarah M. N. Woolley

Note: Based on above, removed L from A1.
'''
cortex_dict = {
    'L3-2b': 'A1',
    'L3': 'A1',
    'L1': 'A1',
    'L2b': 'A1',
    'L2B': 'A1',
    'L2A': 'A1',
    'L2A-L1': 'A1',
    'L3-L2B': 'A1',
    'L3-L2A': 'A1',
    'L2A-L2B': 'A1',
    'L1-L2A': 'A1',
    'CML': 'A1',
    'L1-CML': 'A1',
    'CML-L1': 'A1',
    
    'NCM': 'A2',
    'CMM': 'A2',
}
df['auditory_cortex'] = df['brain_region'].replace(cortex_dict)
df['auditory_cortex'].value_counts()/len(df) * 100

df['auditory_cortex'] = df['brain_region'].map(cortex_dict)
df['auditory_cortex'].value_counts()/len(df) * 100

### WhiWhi: Remove Site5 
```
In WhiWhi, site5 stimuli are not used on other sites (leading to reduced set intersection). This is true for stim_numbers as well as original_wavefiles.
```

df[(df['subject']=='WhiWhi4522M')&(df['site']=='Site1')]['original_wavfile'].unique()

df[(df['subject']=='WhiWhi4522M')&(df['site']=='Site5')]['original_wavfile'].unique()

df[(df['subject']=='WhiWhi4522M')&(df['site']=='Site1')]['original_wavfile'].unique() == df[(df['subject']=='WhiWhi4522M')&(df['site']=='Site2')]['original_wavfile'].unique()

df[(df['subject']=='WhiWhi4522M')&(df['site']=='Site1')]['original_wavfile'].unique() == df[(df['subject']=='WhiWhi4522M')&(df['site']=='Site5')]['original_wavfile'].unique()

df = df[~((df['subject']=='WhiWhi4522M')&(df['site']=='Site5'))]

## Verify and Save

df = df.dropna(subset=['auditory_cortex'])

df.isnull().sum()/len(df)

df.sample(5)

df.to_csv('aa-4.csv', index=False)
